refatoração 'docker_clean.sh'

+++ docker_clean.sh
@@ -8,16 +8,35 @@ then
 	exec sudo bash "$0" "$@"
 fi
 
+
+CONTAINERS="$(docker ps -a -q)"
+IMAGES="$(docker images -a -q)"
+
+
 clear
-echo Parando todos os containers...
-docker stop $(docker ps -a -q)
-sleep 0.5
-echo " "
+if test -z $CONTAINERS
+then
+	echo Não há containers na máquina
+else
+    NUMBER_OF_CONTAINERS="$(docker ps -a -q | wc -l)"
 
-echo Destruindo todos os containers...
-docker rm $(docker ps -a -q)
-sleep 0.5
-echo " "
+	echo Parando todos os containers ativos...
+	docker stop $(docker ps -a -q)
+	sleep 0.5
+	echo " "
 
-echo Destruindo todas as imagens...
-docker rmi $(docker images -a -q)
+	echo "Destruindo todos os $NUMBER_OF_CONTAINERS containers..."
+	docker rm $(docker ps -a -q)
+	sleep 0.5
+	echo " "
+fi
+
+if test -z $IMAGES
+then
+	echo Não há imagens na máquina
+else
+    NUMBER_OF_IMAGES="$(docker images -a -q | wc -l)"
+
+	echo "Destruindo todas as $NUMBER_OF_IMAGES imagens..."
+	docker rmi $(docker images -a -q)
+fi
